refactor: move the query command to internal/cli (#529 slice 4)
Fourth slice of #529: the query command (the largest read command) moves into
internal/cli, following the status pilot's proven lift-and-shift.

- internal/cli/query.go: queryCmd + runQuery + flags moved from cmd/bintrail
  (package main -> cli). The slice-1 DuckDB-tuning calls lose their cli. prefix
  (now same package); bindCommandEnv -> BindCommandEnv; rootCmd.AddCommand
  dropped in favor of AddReadCommands.
- internal/cli/commands.go: AddReadCommands now also registers queryCmd.
- CleanPKList (was cleanPKList in query.go) is exported because recover.go (still
  in cmd/bintrail) shares it; cmd/bintrail/recover.go repointed to cli.CleanPKList
  (it'll de-prefix when recover moves next).
- cmd/bintrail/read_commands_test.go: query added to the wired-on-root guard.
- query_test.go moved; TestQueryCmd_registered adapted to AddReadCommands.

No behavior change. Validated: full unit suite + `go test -tags integration -run
TestEndToEnd` (runs the real `bintrail query`) green. Part of #529.
