+++ internal/cli/commands.go
@@ -8,7 +8,9 @@ import "github.com/spf13/cobra"
 // shim surface without duplicating the command definitions.
 //
 // Commands are migrated into internal/cli one slice at a time (#529); today this
-// registers only status. As the others move, they join this function.
+// registers status and query. As the others (recover, reconstruct, shim) move,
+// they join this function.
 func AddReadCommands(root *cobra.Command) {
 	root.AddCommand(statusCmd)
+	root.AddCommand(queryCmd)
 }
